# openapi.yaml
openapi: 3.0.1
info:
  title: Simple Social Media Application API
  version: 1.0.0
  description: |
    API for a basic social networking service supporting posts, comments, and likes. Designed for MVP and educational use.
servers:
  - url: http://localhost:8080/api
paths:
  /posts:
    get:
      summary: List all posts
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
    post:
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreateRequest'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
  /posts/{postId}:
    get:
      summary: Get a single post
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200':
          description: Post details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update a post
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpdateRequest'
      responses:
        '200':
          description: Post updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a post
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '204':
          description: Post deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /posts/{postId}/comments:
    get:
      summary: List comments for a post
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: Create a comment on a post
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreateRequest'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /posts/{postId}/comments/{commentId}:
    get:
      summary: Get a specific comment
      parameters:
        - $ref: '#/components/parameters/PostId'
        - $ref: '#/components/parameters/CommentId'
      responses:
        '200':
          description: Comment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update a comment
      parameters:
        - $ref: '#/components/parameters/PostId'
        - $ref: '#/components/parameters/CommentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentUpdateRequest'
      responses:
        '200':
          description: Comment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a comment
      parameters:
        - $ref: '#/components/parameters/PostId'
        - $ref: '#/components/parameters/CommentId'
      responses:
        '204':
          description: Comment deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /posts/{postId}/likes:
    post:
      summary: Like a post
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LikeRequest'
      responses:
        '201':
          description: Post liked
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Unlike a post
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LikeRequest'
      responses:
        '204':
          description: Like removed
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    PostId:
      name: postId
      in: path
      required: true
      schema:
        type: string
      description: Unique identifier for the post
    CommentId:
      name: commentId
      in: path
      required: true
      schema:
        type: string
      description: Unique identifier for the comment
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Post:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        likeCount:
          type: integer
        commentCount:
          type: integer
      required: [id, username, content, createdAt, updatedAt, likeCount, commentCount]
    PostCreateRequest:
      type: object
      properties:
        username:
          type: string
        content:
          type: string
      required: [username, content]
    PostUpdateRequest:
      type: object
      properties:
        username:
          type: string
        content:
          type: string
      required: [username, content]
    Comment:
      type: object
      properties:
        id:
          type: string
        postId:
          type: string
        username:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, postId, username, content, createdAt, updatedAt]
    CommentCreateRequest:
      type: object
      properties:
        username:
          type: string
        content:
          type: string
      required: [username, content]
    CommentUpdateRequest:
      type: object
      properties:
        username:
          type: string
        content:
          type: string
      required: [username, content]
    LikeRequest:
      type: object
      properties:
        username:
          type: string
      required: [username]
    Error:
      type: object
      properties:
        message:
          type: string
      required: [message]